`[ob-]broadcast-tx' and `sendtx' don't exist anymore
The documentation and code still made references to these old names.
This commit updates not only these command names, but also any
surrounding text that is relevant.

Maintaining this sort of thing may become increasingly difficult
unless steps are taken to centralize documentation and strings
in a way that can be referenced in multiple places rather than
merely copied.