Update spesmilo/sx/tools/README.md adding ABIS
This adds notes referring to ABIS / ABISprotocol and suggests that rather than specific charities or organizations, that the user be given full discretion to decide how donations should be configured, but leaves much discretion of the amount of the microdonation to the developers of sx based on how that operates in the context of the programming.  In this scenario, the user also has leeway to determine how much and to whom.  This is a pull request for the README.md, so it is conceptual in nature, and does not yet propose to alter the underlying code for sx / tools, but it is hoped that after some discussion that might occur.